resolved warning, thanks to Benjamin!
git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@19842 060e22d8-9d1c-0410-868a-e3a950d63d95

+++ src/drt_acou/acou_ele_calc.cpp
@@ -1542,7 +1542,7 @@ UpdateInteriorVariablesAndComputeResidual(DRT::Discretization &     discretizati
     tempVecv1.Scale(dt*dirk_b[stage]/dirk_a[stage][stage]);
     ele.eleinteriorVelnp_ += tempVecv1;
 
-    if(unsigned(stage)==dirk_q-1)
+    if(stage==dirk_q-1)
     {
       ele.eleinteriorPressn_ = ele.eleinteriorPressnp_;
       ele.eleinteriorVeln_ = ele.eleinteriorVelnp_;
